fix plugin
